backend/requirements/dev.txt
-r base.txt

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-asyncio==0.24.0
factory-boy==3.3.1

# Debug
django-debug-toolbar==4.4.6

# Linting
ruff==0.8.6

# Pre-commit
pre-commit==4.0.1
